fix mission_end handling
On graceful server termination the state file was getting deleted even
if the campaign was not completed. This is because the mission_end event
is fired any time the mission ends. So we need to filter out cases
where the server is restarting and the campaign is not complete.

Closes: #125
Fixes: 903ade6231c7 ("tickets system: add basic ticket accounting system")

+++ src/dct/Theater.lua
@@ -214,9 +214,14 @@ end
 function Theater:onEvent(event)
 	self:notify(event)
 	if event.id == world.event.S_EVENT_MISSION_END then
-		local ok, err = os.remove(settings.statepath)
-		if not ok then
-			Logger:error("unable to remove statepath; "..err)
+		-- Only delete the state if there is an end mission event
+		-- and tickets are complete, otherwise when a server is
+		-- shutdown gracefully the state will be deleted.
+		if self.tickets:isComplete() then
+			local ok, err = os.remove(settings.statepath)
+			if not ok then
+				Logger:error("unable to remove statepath; "..err)
+			end
 		end
 	end
 end
